New project, initialization
Implementation of a Video Graphics Array Controller/Driver. The goal is
to successfully drive a typical monitor and display an image in it. For
the purpose of continues representation through the VGA port, part of
the internal RAM of the FPGA unit used will be assigned as Video RAM
(VRAM) of the driver. The suggested sample image for the testing of the
aforementioned driver is the typical red, blue, green,black horizontal
stripes separated repeatedly by white stripes. The black stripes part is
also repeatedly overlapped by a red,green,blue vertical stripe.